+++ Examples/Resources/CertReq/2-CertReq_RequestSSLCert_Config.ps1
@@ -48,7 +48,7 @@ configuration CertReq_RequestSSLCert_Config
             Subject             = 'foodomain.test.net'
             KeyLength           = '2048'
             Exportable          = $true
-            ProviderName        = '"Microsoft RSA SChannel Cryptographic Provider"'
+            ProviderName        = 'Microsoft RSA SChannel Cryptographic Provider'
             OID                 = '1.3.6.1.5.5.7.3.1'
             KeyUsage            = '0xa0'
             CertificateTemplate = 'WebServer'